requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Data Visualization
plotly>=5.15.0
altair>=5.0.0

# API Integration
requests>=2.31.0
aiohttp>=3.8.0
yfinance>=0.2.18

# Data Processing
python-dotenv>=1.0.0
textblob>=0.17.1

# Database & Storage
sqlite3
sqlalchemy>=2.0.0

# Email & Notifications
smtplib
email

# Utilities
asyncio
threading
logging
json
time
datetime
os
sys
re
warnings
typing
dataclasses
enum

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced Features
# transformers>=4.30.0  # For advanced NLP
# torch>=2.0.0         # For ML features
# scikit-learn>=1.3.0  # For data analysis
